agent/requirements.txt
google-adk
litellm
uvicorn
fastapi